requirements.txt
--extra-index-url https://pypi.fury.io/9sWu_vahYMyGeaBAZ8ac/melizalab/
appnope==0.1.0
backcall==0.1.0
cycler==0.10.0
decorator==4.3.0
-e git+git@gracula.psyc.virginia.edu:pub/dmeliza/dstrf.git@477b8146ab1f4ffb071ade1af6f71daeccfa89a4#egg=dstrf
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
mat-neuron==0.4.1
matplotlib==2.2.2
mpmath==1.0.0
neurofit==0.3.5
numpy==1.14.2
pandas==0.23.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pint==0.8.1
progressbar2==3.38.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pybind11==2.2.3
Pygments==2.2.0
pyparsing==2.2.0
pyspike==0.5.3
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
pyzmq==17.1.0
quickspikes==1.3.4
ruamel.yaml==0.15.44
scipy==1.0.1
seaborn==0.9.0
simplegeneric==0.8.1
six==1.11.0
spyks==0.6.5
sympy==1.2
Theano==1.0.1
tornado==5.1
traitlets==4.3.2
wcwidth==0.1.7
emcee==2.2.1
-e git+https://github.com/tdrobbins/gammatone@fb9390ed9fb249de52227a81283a0db0884c1da2#egg=Gammatone
ipykernel==4.6.1
matplotlib==2.0.2
neurofit==0.3.5
mat-neuron==0.4.1
nose==1.3.7
numpy==1.13.1
scipy==0.19.0
seaborn==0.7.1
progressbar2==3.18.0
pyspike==0.5.1
toelis==2.0.1
ewave==1.0.5
pkgconfig==1.2.2
libtfr==2.0.2
corner==2.0.1
